requirements.txt
avltree==1.1.2
bitarray==2.9.2
blinker==1.7.0
bplustree==0.0.3
BTrees==5.2
cachetools==5.3.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
english-words==2.0.1
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
msgpack==1.0.8
persistent==5.2
pycparser==2.21
pymongo==4.6.2
python-dotenv==1.0.1
rwlock==0.0.7
sortedcontainers==2.4.0
Werkzeug==3.0.1
zope.interface==6.2
